fix: Fixed start and end dates in fawn.
Start and end dates in fawn, if None should be replaced by
real dates.  Start date should be 1998-01-01 (when the FAWN
program started) and end date should be datetime.datetime.now().

Implemented the reportType equal to "entire".  The FAWN web
site doesn't return anything at the moment, but the code
will be there when it is working.